Ainda falta criar o relacionamento entre tabelas do banco de dados para que na hora que o usuário finalize a compra, ela seja salva também no banco de dados.

+++ src/models/Shipping.js
@@ -14,7 +14,7 @@ module.exports = (sequelize, DataType) => {
             allowNull: false
         },
         valor: {
-            type: DataType.INTEGER,
+            type: DataType.DECIMAL,
             allowNull: false
         }
     }, {
